Move Spanner persistence implementation next to the others
